non-functionality code-refactor, prepare for python
