+++ resources/html/AboutDialog.html
@@ -156,6 +156,7 @@ lABPVqVoVsd6gA2YoAVC+QMLSLCBWoV1pWONaxTeygm52rWqnAgCADs=" alt="Donate"></a></p>
     <a href="https://github.com/GilesBathgate">Giles Bathgate</a>,
     <a href="https://github.com/gregjurman">gregjurman</a>,
     <a href="https://github.com/gringer">David Eccles</a>,
+    <a href="https://github.com/gsohler">G&uuml;nther Sohler</a>,
     <a href="https://github.com/henwist">henwist</a>,
     <a href="https://github.com/herbertp">Herbert Poetzl</a>,
     <a href="https://github.com/hoijui">hoijui</a>,
