+++ src/core/LocalScope.cc
@@ -3,7 +3,7 @@
 #include "ModuleInstantiation.h"
 #include "UserModule.h"
 #include "function.h"
-#include "node.h"
+#include "core/node.h"
 
 void LocalScope::addModuleInst(const shared_ptr<ModuleInstantiation>& modinst)
 {
