+++ src/core/RotateExtrudeNode.cc
@@ -30,7 +30,7 @@
 #include "Children.h"
 #include "Parameters.h"
 #include "printutils.h"
-#include "fileutils.h"
+#include "io/fileutils.h"
 #include "Builtins.h"
 #include "handle_dep.h"
 #include <cmath>
