+++ src/core/builtin_functions.cc
@@ -34,8 +34,8 @@
 #include "degree_trig.h"
 #include "FreetypeRenderer.h"
 #include "Parameters.h"
-#include "import.h"
-#include "fileutils.h"
+#include "io/import.h"
+#include "io/fileutils.h"
 
 #include <cmath>
 #include <sstream>
@@ -721,25 +721,6 @@ Value builtin_parent_module(Arguments arguments, const Location& loc)
   return {UserModule::stack_element(s - 1 - n)};
 }
 
-Value builtin_evalstring(const std::shared_ptr<const Context>& context, const FunctionCall * func_call)
-{
-  Arguments arguments(func_call->arguments, context);
-  Location loc = Location::NONE;
-  std::string parname;
-  if (arguments.size() == 0) {
-  } else if (!check_arguments("evalstring", arguments, loc, { Value::Type::STRING })) {
-    return Value::undefined.clone();
-  } else {
-    parname = arguments[0]->toString();
-  }
-
-  boost::optional<const Value&> result = context->try_lookup_variable(parname);
-  if (!result) {
-    return Value::undefined.clone();
-  }
-  return result->clone();
-}
-
 Value builtin_norm(Arguments arguments, const Location& loc)
 {
   if (!check_arguments("norm", arguments, loc, { Value::Type::VECTOR })) {
@@ -1139,11 +1120,6 @@ void register_builtin_functions()
   Builtins::init("parent_module", new BuiltinFunction(&builtin_parent_module),
   {
     "parent_module(number) -> string",
-  });
-
-  Builtins::init("evalstring", new BuiltinFunction(&builtin_evalstring),
-  {
-    "evalstring(string) -> expression",
   });
 
   Builtins::init("is_undef", new BuiltinFunction(&builtin_is_undef),
