+++ src/core/control.cc
@@ -26,7 +26,7 @@
 
 #include "module.h"
 #include "ModuleInstantiation.h"
-#include "node.h"
+#include "core/node.h"
 #include "Arguments.h"
 #include "Children.h"
 #include "Expression.h"
