+++ src/gui/ScadLexer.cc
@@ -19,7 +19,7 @@ ScadLexer::ScadLexer(QObject *parent) : QsciLexerCPP(parent)
     "abs sign rands min max sin cos asin acos tan atan atan2 "
     "round ceil floor pow sqrt exp len log ln str chr ord concat "
     "is_undef is_list is_num is_bool is_string is_function "
-    "lookup search version version_num norm cross parent_module evalstring"
+    "lookup search version version_num norm cross parent_module "
     "dxf_dim dxf_cross";
 
   // -> used in comments only like /*! \cube */
@@ -125,7 +125,7 @@ void Lex::default_rules()
 
   std::string functions("abs sign rands min max sin cos asin acos tan atan atan2 round "
                         "ceil floor pow sqrt exp len log ln str chr ord concat lookup "
-                        "search version version_num norm cross parent_module parent_param dxf_dim "
+                        "search version version_num norm cross parent_module dxf_dim "
                         "dxf_cross is_undef is_list is_num is_bool is_string "
                         "is_function is_object");
   defineRules(functions, efunction);
